+++ Contorls/UserGridViewControl.xaml
@@ -1,11 +1,11 @@
-﻿<UserControl x:Class="BoxNestGroup.Contorls.UserGriviewControl"
+﻿<UserControl x:Class="BoxNestGroup.Contorls.UserGridViewControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:BoxNestGroup"
-             xmlns:view="clr-namespace:BoxNestGroup.View"
-             xmlns:manager="clr-namespace:BoxNestGroup.Manager"
+             xmlns:view="clr-namespace:BoxNestGroup.Views"
+             xmlns:manager="clr-namespace:BoxNestGroup.Managers"
              xmlns:control="clr-namespace:BoxNestGroup.Contorls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
